requirements.txt
Celery
django>=2.0
django-graphql-jwt
django-cors-headers
djangorestframework
docx
python-docx
graphene-django
psycopg2-binary
Scrapy
scrapyd
#git+https://github.com/knil-sama/scrapyd.git@fix-start-urls#egg=scrapyd
python-scrapyd-api
selenium
Spacy
pillow
Pypandoc
pyvirtualdisplay
